MMA-5127 endret varsel tid til å lese fra vault og oppgradert spring …

+++ app/src/main/java/no/nav/dokdistavstemming/consumer/dokumentdistribusjon/HentForsendelseKvitteringIkkeMottatt.java
@@ -11,7 +11,7 @@ import java.util.List;
  */
 public interface HentForsendelseKvitteringIkkeMottatt {
 
-	List<AvstemForsendelseRequestTo> hentForsendelserKvitteringIkkeMottatt(String distribusjonKanal, Long antallTimer);
+	List<AvstemForsendelseRequestTo> hentForsendelserKvitteringIkkeMottatt(String distribusjonKanal, int antallTimer);
 
 	void oppdaterForsendelserAvstemtDatoOgReferanse(OppdaterForsendelserAvstemtInfo oppdaterForsendelserAvstemtInfo);
